% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posixt.R
\name{as-zoned-time-posixt}
\alias{as-zoned-time-posixt}
\alias{as_zoned_time.POSIXt}
\title{Convert to a zoned-time from a date-time}
\usage{
\method{as_zoned_time}{POSIXt}(x, ...)
}
\arguments{
\item{x}{\verb{[POSIXct / POSIXlt]}

A date-time.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A zoned-time.
}
\description{
This is a POSIXct/POSIXlt method for the \code{\link[=as_zoned_time]{as_zoned_time()}} generic.

Converting from one of R's native date-time classes (POSIXct or POSIXlt)
will retain the time zone of that object. There is no \code{zone} argument.
}
\examples{
x <- as.POSIXct("2019-01-01", tz = "America/New_York")
as_zoned_time(x)
}
