% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-calendar.R,
%   R/scale-gregorian-year-month-day.R, R/scale-quarterly-year-quarter-day.R,
%   R/scale-week-year-week-day.R
\name{calendar-scales-color}
\alias{calendar-scales-color}
\alias{scale_colour_year_month_day}
\alias{scale_color_year_month_day}
\alias{scale_fill_year_month_day}
\alias{scale_colour_year_quarter_day}
\alias{scale_color_year_quarter_day}
\alias{scale_fill_year_quarter_day}
\alias{scale_colour_year_week_day}
\alias{scale_color_year_week_day}
\alias{scale_fill_year_week_day}
\title{Gradient color scales: calendar}
\usage{
scale_colour_year_month_day(
  ...,
  low = "#132B43",
  high = "#56B1F7",
  na.value = "grey50",
  guide = "colourbar"
)

scale_color_year_month_day(
  ...,
  low = "#132B43",
  high = "#56B1F7",
  na.value = "grey50",
  guide = "colourbar"
)

scale_fill_year_month_day(
  ...,
  low = "#132B43",
  high = "#56B1F7",
  na.value = "grey50",
  guide = "colourbar"
)

scale_colour_year_quarter_day(
  ...,
  low = "#132B43",
  high = "#56B1F7",
  na.value = "grey50",
  guide = "colourbar"
)

scale_color_year_quarter_day(
  ...,
  low = "#132B43",
  high = "#56B1F7",
  na.value = "grey50",
  guide = "colourbar"
)

scale_fill_year_quarter_day(
  ...,
  low = "#132B43",
  high = "#56B1F7",
  na.value = "grey50",
  guide = "colourbar"
)

scale_colour_year_week_day(
  ...,
  low = "#132B43",
  high = "#56B1F7",
  na.value = "grey50",
  guide = "colourbar"
)

scale_color_year_week_day(
  ...,
  low = "#132B43",
  high = "#56B1F7",
  na.value = "grey50",
  guide = "colourbar"
)

scale_fill_year_week_day(
  ...,
  low = "#132B43",
  high = "#56B1F7",
  na.value = "grey50",
  guide = "colourbar"
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{low, high}{Colours for low and high ends of the gradient.}

\item{na.value}{Colour to use for missing values}

\item{guide}{Type of legend. Use \code{"colourbar"} for continuous
colour bar, or \code{"legend"} for discrete colour legend.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Gradient color scales for use with ggplot2.
\itemize{
\item \code{scale_color_year_month_day()} and \code{scale_fill_year_month_day()} are only
valid on year and month precision inputs.
\item \code{scale_color_year_quarter_day()} and \code{scale_fill_year_quarter_day()} are
only valid on year and quarter precision inputs.
\item \code{scale_color_year_week_day()} and \code{scale_fill_year_week_day()} are only
valid on year and week precision inputs.
}

For day precision and finer, we currently still recommend using Date and
POSIXct with \code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_date()}} and
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_datetime()}}.
}
\examples{
\dontshow{if (rlang::is_installed("ggplot2") && rlang::is_installed("scales")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(ggplot2)
library(vctrs)

# ---------------------------------------------------------------------------
# Color

set.seed(123)

date <- year_quarter_day(2019, 1) + duration_quarters(1:100)
height <- ifelse(
  date < year_quarter_day(2033, 1),
  sample(30:50, size = 50, replace = TRUE),
  sample(40:60, size = 50, replace = TRUE)
)
weight <- ifelse(
  date < year_quarter_day(2028, 1),
  sample(100:150, size = 50, replace = TRUE),
  sample(100:180, size = 50, replace = TRUE)
)

df <- data_frame(
  date = date,
  height = height,
  weight = weight
)

ggplot(df, aes(height, weight, color = date)) +
  geom_point() +
  scale_colour_year_quarter_day(
    low = "red",
    high = "blue"
  )

# ---------------------------------------------------------------------------
# Fill

economics$date <- as_year_month_day(economics$date)
economics$date <- calendar_narrow(economics$date, "month")

ggplot(economics, aes(x = date, y = unemploy, fill = date)) +
  geom_col()
\dontshow{\}) # examplesIf}
}
