% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-calendar.R,
%   R/scale-gregorian-year-month-day.R, R/scale-quarterly-year-quarter-day.R,
%   R/scale-week-year-week-day.R
\name{calendar-scales-size}
\alias{calendar-scales-size}
\alias{scale_size_year_month_day}
\alias{scale_size_year_quarter_day}
\alias{scale_size_year_week_day}
\title{Size scales: calendar}
\usage{
scale_size_year_month_day(..., range = c(1, 6))

scale_size_year_quarter_day(..., range = c(1, 6))

scale_size_year_week_day(..., range = c(1, 6))
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{range}{a numeric vector of length 2 that specifies the minimum and
maximum size of the plotting symbol after transformation.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Size scales for use with ggplot2.
\itemize{
\item \code{scale_size_year_month_day()} is only valid on year and month precision
inputs.
\item \code{scale_size_year_quarter_day()} is only valid on year and quarter
precision inputs.
\item \code{scale_size_year_week_day()} is only valid on year and week precision
inputs.
}

For day precision and finer, we currently still recommend using Date and
POSIXct with \code{\link[ggplot2:scale_size]{ggplot2::scale_size_date()}} and
\code{\link[ggplot2:scale_size]{ggplot2::scale_size_datetime()}}.
}
\examples{
\dontshow{if (rlang::is_installed("ggplot2") && rlang::is_installed("scales")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(ggplot2)
library(vctrs)

set.seed(123)

date <- year_quarter_day(2019, 1) + duration_quarters(1:100)
height <- ifelse(
  date < year_quarter_day(2033, 1),
  sample(30:50, size = 50, replace = TRUE),
  sample(40:60, size = 50, replace = TRUE)
)
weight <- ifelse(
  date < year_quarter_day(2028, 1),
  sample(100:150, size = 50, replace = TRUE),
  sample(100:180, size = 50, replace = TRUE)
)

df <- data_frame(
  date = date,
  height = height,
  weight = weight
)

ggplot(df, aes(height, weight, size = date)) +
  geom_point() +
  scale_size_year_quarter_day(range = c(1, 10))
\dontshow{\}) # examplesIf}
}
