% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendar_group}
\alias{calendar_group}
\title{Group calendar components}
\usage{
calendar_group(x, precision, ..., n = 1L)
}
\arguments{
\item{x}{\verb{[calendar]}

A calendar vector.}

\item{precision}{\verb{[character(1)]}

A precision. Allowed precisions are dependent on the calendar used.}

\item{...}{These dots are for future extensions and must be empty.}

\item{n}{\verb{[positive integer(1)]}

A single positive integer specifying a multiple of \code{precision} to
group by.}
}
\value{
\code{x} grouped at the specified \code{precision}.
}
\description{
\code{calendar_group()} groups calendar vectors to a multiple of the
specified precision. The value of the component that matches the supplied
\code{precision} will be adjusted, and more precise components will be dropped.
}
\examples{
x <- year_month_day(2019, c(1, 1, 2, 2, 3, 3, 4, 4), 1:8)
x

# Group by two months
calendar_group(x, "month", n = 2)

# Group by two days of the month
calendar_group(x, "day", n = 2)

y <- c(year_quarter_day(2019, 1, 50:60), year_quarter_day(2019, 2, 50:60))

# Group by 3 days of the current quarter
calendar_group(y, "day", n = 3)

z <- c(iso_year_week_day(2019, 1:10), iso_year_week_day(2020, 1:10))

# Group by 5 ISO weeks of the current ISO year
calendar_group(z, "week", n = 5)
}
