% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendar_narrow}
\alias{calendar_narrow}
\title{Narrow a calendar to a less precise precision}
\usage{
calendar_narrow(x, precision)
}
\arguments{
\item{x}{\verb{[calendar]}

A calendar vector.}

\item{precision}{\verb{[character(1)]}

A precision. Allowed precisions are dependent on the calendar used.}
}
\value{
\code{x} narrowed to the supplied \code{precision}.
}
\description{
\code{calendar_narrow()} drops \code{x} to the specified \code{precision}.
}
\examples{
# Second precision
x <- year_month_day(2019, 1, 1, 1, 1, 1)
x

# Drop to day precision
calendar_narrow(x, "day")
}
