% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendar_precision}
\alias{calendar_precision}
\title{Precision: calendar}
\usage{
calendar_precision(x)
}
\arguments{
\item{x}{\verb{[clock_calendar]}

A calendar.}
}
\value{
A single string holding the precision of the calendar.
}
\description{
\code{calendar_precision()} extracts the precision from a calendar object. It
returns the precision as a single string.
}
\examples{
calendar_precision(year_month_day(2019))
calendar_precision(year_month_day(2019, 1, 1))
calendar_precision(year_quarter_day(2019, 3))
}
