% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civil-package.R
\docType{package}
\name{civil-package}
\alias{civil}
\alias{civil-package}
\title{civil: Local and Zoned Dates}
\description{
Provides comprehensive support for local and zoned
    dates and times. A local date does not assume any particular time zone.
    A zoned date is a local date with a time zone attached. Precision up
    to nanoseconds is supported, and partial date types representing
    coarser periods such as months or weeks are also implemented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/civil}
  \item Report bugs at \url{https://github.com/DavisVaughan/civil/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
