% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian-year-month-day.R,
%   R/gregorian-year-month-weekday.R
\name{clock-arith}
\alias{clock-arith}
\alias{vec_arith.clock_year_month_day}
\alias{vec_arith.clock_year_month_weekday}
\title{Support for vctrs arithmetic}
\usage{
\method{vec_arith}{clock_year_month_day}(op, x, y, ...)

\method{vec_arith}{clock_year_month_weekday}(op, x, y, ...)
}
\arguments{
\item{op}{An arithmetic operator as a string}

\item{x}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{y}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
The result of the arithmetic operation.
}
\description{
Support for vctrs arithmetic
}
\examples{
vctrs::vec_arith("+", year_month_day(2019), 1)
}
