% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock-codes.R
\docType{data}
\name{clock_codes}
\alias{clock_codes}
\title{Integer codes}
\format{
An environment containing mappings from month names and weekday names
to integer codes.
}
\usage{
clock_codes
}
\description{
A locked environment with useful mappings from month names and weekday names
to integer codes.
\subsection{Month codes}{
\itemize{
\item \code{january == 1}
\item \code{february == 2}
\item \code{march == 3}
\item \code{april == 4}
\item \code{may == 5}
\item \code{june == 6}
\item \code{july == 7}
\item \code{august == 8}
\item \code{september == 9}
\item \code{october == 10}
\item \code{november == 11}
\item \code{december == 12}
}
}

\subsection{Weekday codes}{
\itemize{
\item \code{sunday == 1}
\item \code{monday == 2}
\item \code{tuesday == 3}
\item \code{wednesday == 4}
\item \code{thursday == 5}
\item \code{friday == 6}
\item \code{saturday == 7}
}
}

\subsection{ISO weekday codes}{

In \code{clock_codes$iso}:
\itemize{
\item \code{monday == 1}
\item \code{tuesday == 2}
\item \code{wednesday == 3}
\item \code{thursday == 4}
\item \code{friday == 5}
\item \code{saturday == 6}
\item \code{sunday == 7}
}
}
}
\examples{
weekday(clock_codes$wednesday)

year_month_weekday(2019, clock_codes$april, clock_codes$monday, 1:4)

iso_year_week_day(2020, 52, clock_codes$iso$thursday)
}
\keyword{datasets}
