% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock-mapping.R
\name{clock_mapping}
\alias{clock_mapping}
\alias{clock_mapping_lookup}
\alias{clock_mapping_languages}
\title{Create or retrieve date related mappings}
\usage{
clock_mapping(mon, mon_ab = mon, day, day_ab = day, am_pm = c("AM", "PM"))

clock_mapping_lookup(language)

clock_mapping_languages()
}
\arguments{
\item{mon, mon_ab}{\verb{[character(12)]}

Full and abbreviated month names. Starts with January.}

\item{day, day_ab}{\verb{[character(7)]}

Full and abbreviated week day names. Starts with Sunday.}

\item{am_pm}{\verb{[character(2)]}

Names used for AM and PM.}

\item{language}{\verb{[character(1)]}

A BCP 47 locale, generally constructed from a two or three
digit language code. See \code{clock_mapping_languages()} for a complete list of
available locales.}
}
\description{
When parsing and formatting dates, you often need to know how weekdays of
the week and months are represented as text. These functions allow you
to either create your own mapping, or look one up from a standard set of
language specific mappings. The standard list is derived from ICU
(\url{https://unicode-org.github.io/icu/}) via the stringi package.
\itemize{
\item \code{clock_mapping_lookup()} looks up a set of mapping information from a
given language code.
\item \code{clock_mapping_languages()} lists the language codes that are accepted.
\item \code{clock_mapping()} lets you create your own mapping. Use this if the
currently supported languages don't meet your needs.
}
}
\examples{
clock_mapping_lookup("en")
clock_mapping_lookup("ko")
clock_mapping_lookup("fr")
}
