% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-locale.R
\name{date_locale}
\alias{date_locale}
\alias{default_date_locale}
\title{Create a date locale}
\usage{
date_locale(date_names = "en", decimal_mark = ".")

default_date_locale()
}
\arguments{
\item{date_names}{Character representations of localized weekday names,
month names, and AM/PM names. Either the language code as string
(passed on to \code{\link[=date_names_lang]{date_names_lang()}}) or an object created by \code{\link[=date_names]{date_names()}}.}

\item{decimal_mark}{Symbol used for the decimal place when formatting
sub-second date-times. Either \code{","} or \code{"."}.}
}
\description{
A date locale contains the information required to format and parse dates and
date times. The defaults have been chosen to match US English. A date locale
object can be provided to \code{format()} methods to override the defaults.
Alternatively, set the \code{clock.default_date_locale} global option to your
preferred date locale, and that will be used as the default instead.
}
\examples{
date_locale()
date_locale(date_names = "fr")
default_date_locale()
}
