% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{date_weekday_factor}
\alias{date_weekday_factor}
\title{Convert a date or date-time to a weekday factor}
\usage{
date_weekday_factor(
  x,
  ...,
  labels = "en",
  abbreviate = TRUE,
  encoding = "western"
)
}
\arguments{
\item{x}{\verb{[Date / POSIXct / POSIXlt]}

A date or date-time vector.}

\item{...}{These dots are for future extensions and must be empty.}

\item{labels}{\verb{[clock_labels / character(1)]}

Character representations of localized weekday names, month names, and
AM/PM names. Either the language code as string (passed on to
\code{\link[=clock_labels_lookup]{clock_labels_lookup()}}), or an object created by \code{\link[=clock_labels]{clock_labels()}}.}

\item{abbreviate}{\verb{[logical(1)]}

If \code{TRUE}, the abbreviated weekday names from \code{labels} will be used.

If \code{FALSE}, the full weekday names from \code{labels} will be used.}

\item{encoding}{\verb{[character(1)]}

One of:
\itemize{
\item \code{"western"}: Encode the weekdays as an ordered factor with levels from
Sunday -> Saturday.
\item \code{"iso"}: Encode the weekdays as an ordered factor with levels from
Monday -> Sunday.
}}
}
\value{
An ordered factor representing the weekdays.
}
\description{
\code{date_weekday_factor()} converts a date or date-time to an ordered factor
with levels representing the weekday. This can be useful in combination with
ggplot2, or for modeling.
}
\examples{
x <- as.Date("2019-01-01") + 0:6

# Default to Sunday -> Saturday
date_weekday_factor(x)

# ISO encoding is Monday -> Sunday
date_weekday_factor(x, encoding = "iso")

# With full names
date_weekday_factor(x, abbreviate = FALSE)

# Or a different language
date_weekday_factor(x, labels = "fr")
}
