% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{duration-helper}
\alias{duration-helper}
\alias{duration_years}
\alias{duration_quarters}
\alias{duration_months}
\alias{duration_weeks}
\alias{duration_days}
\alias{duration_hours}
\alias{duration_minutes}
\alias{duration_seconds}
\alias{duration_milliseconds}
\alias{duration_microseconds}
\alias{duration_nanoseconds}
\title{Construct a duration}
\usage{
duration_years(n = integer())

duration_quarters(n = integer())

duration_months(n = integer())

duration_weeks(n = integer())

duration_days(n = integer())

duration_hours(n = integer())

duration_minutes(n = integer())

duration_seconds(n = integer())

duration_milliseconds(n = integer())

duration_microseconds(n = integer())

duration_nanoseconds(n = integer())
}
\arguments{
\item{n}{\verb{[integer]}

The number of ticks to use when creating the duration.}
}
\value{
A duration of the specified precision.
}
\description{
These helpers construct durations of the specified precision.
}
\examples{
duration_years(1:5)
duration_nanoseconds(1:5)
}
