% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{duration_precision}
\alias{duration_precision}
\title{Precision: duration}
\usage{
duration_precision(x)
}
\arguments{
\item{x}{\verb{[clock_duration]}

A duration.}
}
\value{
A single string holding the precision of the duration.
}
\description{
\code{duration_precision()} extracts the precision from a duration object. It
returns the precision as a single string.
}
\examples{
duration_precision(duration_seconds(1))
duration_precision(duration_nanoseconds(2))
duration_precision(duration_quarters(1:5))
}
