% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{is_duration}
\alias{is_duration}
\title{Is \code{x} a duration?}
\usage{
is_duration(x)
}
\arguments{
\item{x}{\verb{[object]}

An object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"clock_duration"}, otherwise \code{FALSE}.
}
\description{
This function determines if the input is a duration object.
}
\examples{
is_duration(1)
is_duration(duration_days(1))
}
