% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-time.R
\name{is_sys}
\alias{is_sys}
\title{Is \code{x} a sys-time?}
\usage{
is_sys(x)
}
\arguments{
\item{x}{\verb{[object]}

An object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"clock_sys_time"}, otherwise \code{FALSE}.
}
\description{
This function determines if the input is a sys-time object.
}
\examples{
is_sys(1)
is_sys(as_sys_time(duration_days(1)))
}
