% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekday.R
\name{is_weekday}
\alias{is_weekday}
\title{Is \code{x} a weekday?}
\usage{
is_weekday(x)
}
\arguments{
\item{x}{\verb{[object]}

An object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"clock_weekday"}, otherwise \code{FALSE}.
}
\description{
This function determines if the input is a weekday object.
}
\examples{
is_weekday(1)
is_weekday(weekday(1))
}
