% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian-year-month-day.R
\name{is_year_month_day}
\alias{is_year_month_day}
\title{Is \code{x} a year-month-day?}
\usage{
is_year_month_day(x)
}
\arguments{
\item{x}{\verb{[object]}

An object.}
}
\value{
Returns \code{TRUE} if \code{x} inherits from \code{"clock_year_month_day"},
otherwise returns \code{FALSE}.
}
\description{
Check if \code{x} is a year-month-day.
}
\examples{
is_year_month_day(year_month_day(2019))
}
