% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian-year-month-weekday.R
\name{is_year_month_weekday}
\alias{is_year_month_weekday}
\title{Is \code{x} a year-month-weekday?}
\usage{
is_year_month_weekday(x)
}
\arguments{
\item{x}{\verb{[object]}

An object.}
}
\value{
Returns \code{TRUE} if \code{x} inherits from \code{"clock_year_month_weekday"},
otherwise returns \code{FALSE}.
}
\description{
Check if \code{x} is a year-month-weekday.
}
\examples{
is_year_month_weekday(year_month_weekday(2019))
}
