% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarterly-year-quarter-day.R
\name{is_year_quarter_day}
\alias{is_year_quarter_day}
\title{Is \code{x} a year-quarter-day?}
\usage{
is_year_quarter_day(x)
}
\arguments{
\item{x}{\verb{[object]}

An object.}
}
\value{
Returns \code{TRUE} if \code{x} inherits from \code{"clock_year_quarter_day"},
otherwise returns \code{FALSE}.
}
\description{
Check if \code{x} is a year-quarter-day.
}
\examples{
is_year_quarter_day(year_quarter_day(2019))
}
