% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoned-time.R
\name{is_zoned}
\alias{is_zoned}
\title{Is \code{x} a zoned-time?}
\usage{
is_zoned(x)
}
\arguments{
\item{x}{\verb{[object]}

An object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"clock_zoned_time"}, otherwise \code{FALSE}.
}
\description{
This function determines if the input is a zoned-time object.
}
\examples{
is_zoned(1)
is_zoned(zoned_now("America/New_York"))
}
