% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-time.R
\name{sys_time_now}
\alias{sys_time_now}
\title{What is the current sys-time?}
\usage{
sys_time_now()
}
\value{
A sys-time of the current time in UTC.
}
\description{
\code{sys_time_now()} returns the current time in UTC.
}
\details{
The time is returned with a nanosecond precision, but the actual amount
of data returned is OS dependent. Usually, information at at least the
microsecond level is returned, with some platforms returning nanosecond
information.
}
\examples{
x <- sys_time_now()
}
