% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-point.R
\name{time_point_precision}
\alias{time_point_precision}
\title{Precision: time point}
\usage{
time_point_precision(x)
}
\arguments{
\item{x}{\verb{[clock_time_point]}

A time point.}
}
\value{
A single string holding the precision of the time point.
}
\description{
\code{time_point_precision()} extracts the precision from a time point, such
as a sys-time or naive-time. It returns the precision as a single string.
}
\examples{
time_point_precision(sys_time_now())
time_point_precision(as_naive_time(duration_days(1)))
}
