% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{tkns}
\alias{tkns}
\title{Combine format tokens together}
\usage{
tkns(..., sep = "")
}
\arguments{
\item{...}{Tokens or formats to combine. Can also contain simple strings.}

\item{sep}{Separator to use between tokens.}
}
\description{
\code{tkns()} provides a simple way to combine multiple tokens or formats together
into a new format.
}
\examples{
tkns(tkn_year(), tkn_month(), tkn_day(), sep = "-")
}
