% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekday.R
\name{weekday-encoding}
\alias{weekday-encoding}
\alias{weekday_c_encoding}
\alias{weekday_iso_encoding}
\title{Extract weekday encodings}
\usage{
weekday_c_encoding(x)

weekday_iso_encoding(x)
}
\arguments{
\item{x}{\verb{[weekday]}

A weekday vector.}
}
\value{
An integer vector of encodings.
}
\description{
Given a weekday object, \code{x}:
\itemize{
\item \code{weekday_c_encoding()} returns the underlying C encoding, with a range
of \verb{[0, 6]}, where 0 = Sunday, and 6 = Saturday.
\item \code{weekday_iso_encoding()} returns the underlying ISO encoding, with a
range of \verb{[1, 7]}, where 1 = Monday, and 7 = Sunday.
}
}
\examples{
# Here we supply a C encoding to `weekday()`
x <- weekday(0:6)
x

# Notice that Sunday is the only day that is different
weekday_c_encoding(x)
weekday_iso_encoding(x)
}
