% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekday.R
\name{weekday_encoding}
\alias{weekday_encoding}
\title{Extract weekday encodings}
\usage{
weekday_encoding(x, ..., encoding = "c")
}
\arguments{
\item{x}{\verb{[weekday]}

A weekday vector.}

\item{...}{These dots are for future extensions and must be empty.}

\item{encoding}{\verb{[character(1)]}

One of:
\itemize{
\item \code{"c"}: Encode weekdays with a range of \verb{[0, 6]}, where
0 = Sunday, and 6 = Saturday.
\item \code{"iso"}: Encode weekdays with a range of \verb{[1, 7]}, where
1 = Monday, and 7 = Sunday.
}}
}
\value{
An integer vector of encodings.
}
\description{
\code{weekday_encoding()} extracts out an integer encoding for the weekday.
}
\examples{
# Here we supply a C encoding to `weekday()`
x <- weekday(0:6)
x

# Notice that Sunday is the only day that is different
weekday_encoding(x, encoding = "c")
weekday_encoding(x, encoding = "iso")
}
