% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarterly-year-quarter-day.R
\name{year_quarter_day}
\alias{year_quarter_day}
\title{Calendar: year-quarter-day}
\usage{
year_quarter_day(
  year,
  quarter = NULL,
  day = NULL,
  hour = NULL,
  minute = NULL,
  second = NULL,
  subsecond = NULL,
  ...,
  start = 1L,
  subsecond_precision = NULL
)
}
\arguments{
\item{year}{\verb{[integer]}

The fiscal year. Values \verb{[-9999, 9999]} are allowed.}

\item{quarter}{\verb{[integer / NULL]}

The fiscal quarter. Values \verb{[1, 4]} are allowed.}

\item{day}{\verb{[integer / "last" / NULL]}

The day of the quarter. Values \verb{[1, 92]} are allowed.

If \code{"last"}, the last day of the quarter is returned.}

\item{...}{These dots are for future extensions and must be empty.}

\item{start}{\verb{[integer(1)]}

The month to start the fiscal year in. 1 = January and 12 = December.}
}
\value{
A year-quarter-day calendar vector.
}
\description{
\code{year_quarter_day()} constructs a calendar from the fiscal year, fiscal
quarter, and day of the quarter, along with a value determining which
month the fiscal year \code{start}s in.
}
\details{
Fields are recycled against each other.

Fields are collected in order until the first \code{NULL} field is located. No
fields after the first \code{NULL} field are used.
}
\examples{
# Year-quarter type
x <- year_quarter_day(2019, 1:4)
x

add_quarters(x, 2)

# Set the day to the last day of the quarter
x <- set_day(x, "last")
x

# Start the fiscal year in June
june <- 6L
y <- year_quarter_day(2019, 1:4, "last", start = june)

# Compare the year-month-day values that result from having different
# fiscal year start months
as_year_month_day(x)
as_year_month_day(y)
}
