% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{zone-database}
\alias{zone-database}
\alias{zone_database_version}
\alias{zone_database_names}
\title{Time zone database}
\usage{
zone_database_version()

zone_database_names()
}
\value{
\itemize{
\item \code{zone_database_version()} returns a single string of the database version.
\item \code{zone_database_names()} returns a character vector of zone names.
}
}
\description{
\itemize{
\item \code{zone_database_version()} returns the version of the time zone database
currently in use by clock.
\item \code{zone_database_names()} returns the time zone names found in the database.
}
}
\examples{
zone_database_version()
zone_database_names()
}
