% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoned-time.R
\name{zoned-zone}
\alias{zoned-zone}
\alias{zoned_zone}
\alias{zoned_set_zone}
\title{Retrieve or set the time zone}
\usage{
zoned_zone(x)

zoned_set_zone(x, zone)
}
\arguments{
\item{x}{\verb{[zoned_time / Date / POSIXt]}

A zoned time to get or set the time zone of.}

\item{zone}{\verb{[character(1)]}

A valid time zone to switch to.}
}
\value{
\code{zoned_zone()} returns a string containing the time zone.

\code{zoned_set_zone()} returns \code{x} with an altered time zone attribute. The
underlying instant is \emph{not} changed.
}
\description{
\code{zoned_zone()} retrieves the time zone.

\code{zoned_set_zone()} sets the time zone \emph{without changing the
underlying instant}. This means that the result will represent the equivalent
time in the new time zone.
}
\examples{
x <- as.POSIXct("2019-01-01 00:00:00", tz = "America/New_York")

zoned_zone(x)

# Equivalent UTC time
zoned_set_zone(x, "UTC")

y <- as_zoned_time(x)

# With a zoned-time
zoned_set_zone(y, "UTC")

# To force a new time zone with the same wall time,
# convert to a naive time that has no implied time zone,
# then convert back to a zoned time in the new time zone.
nt <- as_naive_time(y)
nt
as_zoned_time(nt, "UTC")
}
