% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoned-time.R
\name{zoned_dst}
\alias{zoned_dst}
\title{Is daylight saving time active?}
\usage{
zoned_dst(x)
}
\arguments{
\item{x}{\verb{[zoned_time / Date / POSIXt]}

A zoned time to check if daylight saving time is active for.}
}
\value{
A logical the same size as \code{x}. Returns \code{TRUE} if daylight saving
time is active, \code{FALSE} if not, and \code{NA} if \code{x} is \code{NA}.
}
\description{
\code{zoned_dst()} detects if daylight saving time is currently active.
}
\examples{
# R defines Date as UTC, where DST is never active
zoned_dst(as.Date("2019-01-01"))

x <- year_month_day(2021, 03, 14, hour = c(1, 3))
x <- as_naive_time(x)
x <- as_zoned_time(x, "America/New_York")

# Daylight savings time was turned on at 2am
zoned_dst(x)
}
