% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoned-time.R
\name{zoned_info}
\alias{zoned_info}
\title{Retrieve low-level zoned time information}
\usage{
zoned_info(x)
}
\arguments{
\item{x}{\verb{[zoned_time]}

A zoned time.}
}
\value{
A data frame of low level information to retrieve information for.
}
\description{
\code{zoned_info()} retrieves a set of low-level information generally not
required for most date-time manipulations. It returns a data frame with
the following columns:
\itemize{
\item \code{begin}, \code{end}: Zoned times specifying the range of the current daylight
saving time rule. The range is a half-open interval of \verb{[begin, end)}.
\item \code{offset}: A second precision \code{duration} specifying the offset from UTC.
\item \code{dst}: A logical vector specifying if daylight saving time is currently
active.
\item \code{abbreviation}: The time zone abbreviation in use throughout
this \code{begin} to \code{end} range.
}
}
\details{
If there have never been any daylight saving time transitions, the minimum
supported year value is returned for \code{begin} (typically, a year value of
\code{-32767}).

If daylight saving time is no longer used in a time zone, the maximum
supported year value is returned for \code{end} (typically, a year value of
\code{32767}).
}
\examples{
x <- year_month_day(2021, 03, 14, c(01, 03), c(59, 00), c(59, 00))
x <- as_naive_time(x)
x <- as_zoned(x, "America/New_York")

# x[1] is in EST, x[2] is in EDT
x

info <- zoned_info(x)
info

# `end` can be used to iterate through daylight saving time transitions
# by repeatedly calling `zoned_info()`
zoned_info(info$end)
}
