% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoned-time.R
\name{zoned_offset}
\alias{zoned_offset}
\title{Get the offset from UTC}
\usage{
zoned_offset(x)
}
\arguments{
\item{x}{\verb{[zoned_time / Date / POSIXt]}

A zoned time to extract the offset for.}
}
\description{
\code{zoned_offset()} returns the offset from UTC as a duration of seconds.
}
\examples{
# R defines Date as UTC
zoned_offset(as.Date("2019-01-01"))

x <- year_month_day(2021, 03, 14, hour = c(1, 3))
x <- as_naive_time(x)
x <- as_zoned_time(x, "America/New_York")

# Daylight savings time alters the offset from UTC
zoned_offset(x)

# Can extract directly from POSIXct if you started with one of those
zoned_offset(as.POSIXct(x))
}
