% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{civil-arithmetic}
\alias{civil-arithmetic}
\alias{add_years}
\alias{subtract_years}
\alias{add_months}
\alias{subtract_months}
\alias{add_years_and_months}
\alias{subtract_years_and_months}
\alias{add_weeks}
\alias{subtract_weeks}
\alias{add_days}
\alias{subtract_days}
\alias{add_hours}
\alias{subtract_hours}
\alias{add_minutes}
\alias{subtract_minutes}
\alias{add_seconds}
\alias{subtract_seconds}
\title{Date-time arithmetic}
\usage{
add_years(x, n, ...)

subtract_years(x, n, ...)

add_months(x, n, ...)

subtract_months(x, n, ...)

add_years_and_months(x, years, months, ...)

subtract_years_and_months(x, years, months, ...)

add_weeks(x, n, ...)

subtract_weeks(x, n, ...)

add_days(x, n, ...)

subtract_days(x, n, ...)

add_hours(x, n, ...)

subtract_hours(x, n, ...)

add_minutes(x, n, ...)

subtract_minutes(x, n, ...)

add_seconds(x, n, ...)

subtract_seconds(x, n, ...)
}
\arguments{
\item{x}{\verb{[Date / POSIXct / POSIXlt / local_date / local_datetime]}

A zoned or local date-time vector.}

\item{n}{\verb{[integer]}

An integer vector representing the number of units to add to or
subtract from \code{x}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Date-time arithmetic in civil is broken down into two types: zoned and local.
Each type is documented on its own help page.

\link[=civil-zoned-arithmetic]{Zoned} dates have a time zone attached. The two
base R classes, POSIXct and Date, implement zoned dates (Date is assumed to
implicitly have a time zone of UTC).

\link[=civil-local-arithmetic]{Local} dates are independent of a time zone. The
two civil objects, local_date and local_datetime, implement local dates.
}
