% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civil-package.R
\docType{package}
\name{civil-package}
\alias{civil}
\alias{civil-package}
\title{civil: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/civil}
  \item Report bugs at \url{https://github.com/DavisVaughan/civil/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
