% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock-package.R
\docType{package}
\name{clock-package}
\alias{clock}
\alias{clock-package}
\title{clock: Calendars, Time Points, and Durations}
\description{
Provides a comprehensive interface for date and datetime
    manipulations. Both naive and zoned datetimes are supported, with
    precision up to nanoseconds. Additionally, multiple types of calendars
    have been implemented, allowing for various ways to represent calendar
    dates, such as year-month-day, year-month-weekday, or
    year-quarter-day.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/clock}
  \item Report bugs at \url{https://github.com/DavisVaughan/clock/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
