% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock-package.R
\docType{package}
\name{clock-package}
\alias{clock}
\alias{clock-package}
\title{clock: A Date-Time Library}
\description{
Provides a comprehensive library for date-time manipulations
    using a new family of orthogonal date-time classes (durations, time
    points, zoned times, and calendars) that partition responsibilities so
    that the complexities of time zones are only considered when they are
    really needed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/clock}
  \item Report bugs at \url{https://github.com/r-lib/clock/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
