% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock-package.R
\docType{package}
\name{clock-package}
\alias{clock}
\alias{clock-package}
\title{clock: Date-Time Types and Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a comprehensive library for date-time manipulations using a new family of orthogonal date-time classes (durations, time points, zoned-times, and calendars) that partition responsibilities so that the complexities of time zones are only considered when they are really needed. Capabilities include: date-time parsing, formatting, arithmetic, extraction and updating of components, and rounding.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://clock.r-lib.org}
  \item \url{https://github.com/r-lib/clock}
  \item Report bugs at \url{https://github.com/r-lib/clock/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
