% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian-year-month-day.R
\name{parse_year_month_day}
\alias{parse_year_month_day}
\title{Parse into a year-month-day calendar}
\usage{
parse_year_month_day(
  x,
  ...,
  format = NULL,
  precision = "day",
  locale = default_date_locale()
)
}
\arguments{
\item{x}{\verb{[character]}

A character vector to parse.}

\item{...}{These dots are for future extensions and must be empty.}

\item{format}{\verb{[NULL / character(1)]}

A character format string.

If \code{NULL}, a default format string is chosen based on the \code{precision}.
The default format string is chosen to parse the result of calling
\code{format()} on a year-month-day object. For example, with
\code{precision = "month"}, \code{format} would be set to \verb{"\%Y-\%m}, and with
\code{precision = "millisecond"} it would be set to \code{"\%Y-\%m-\%d \%H:\%M:\%S"}.}

\item{precision}{\verb{[character(1)]}

The precision of the resulting year-month-day.}

\item{locale}{\verb{[clock_locale]}

A locale object created by \code{\link[=date_locale]{date_locale()}}. Defaults to
\code{\link[=default_date_locale]{default_date_locale()}}.}
}
\value{
A year-month-day calendar vector. If a parsing fails, \code{NA} is
returned.
}
\description{
\code{parse_year_month_day()} parses a character vector into a year-month-day
calendar.
}
\examples{
x <- "2019-01-01"

# Default parses at day precision
parse_year_month_day(x)

# Can parse at less precise precisions too
parse_year_month_day(x, precision = "month")
parse_year_month_day(x, precision = "year")

# Even invalid dates can be round-tripped through format<->parse calls
invalid <- year_month_day(2019, 2, 30)
parse_year_month_day(format(invalid))

# Can parse with time of day
x <- year_month_day(
  2019, 01, 30, 02, 30, 00, 5000,
  subsecond_precision = "nanosecond"
)

parse_year_month_day(format(x), precision = "nanosecond")

# Can parse using other format tokens as well
parse_year_month_day(
  "January, 2019",
  format = "\%B, \%Y",
  precision = "month"
)

# Parsing a French year-month-day
parse_year_month_day(
  "octobre 1, 2000",
  format = "\%B \%d, \%Y",
  locale = date_locale("fr")
)
}
