% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzdb.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tzdb_names}
\alias{tzdb_version}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tzdb}{\code{\link[tzdb]{tzdb_names}}, \code{\link[tzdb]{tzdb_version}}}
}}

