% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekday.R
\name{weekday_encoding}
\alias{weekday_encoding}
\title{Extract weekday encodings}
\usage{
weekday_encoding(x, ..., encoding = "western")
}
\arguments{
\item{x}{\verb{[weekday]}

A weekday vector.}

\item{...}{These dots are for future extensions and must be empty.}

\item{encoding}{\verb{[character(1)]}

One of:
\itemize{
\item \code{"western"}: Encode weekdays assuming \code{1 == Sunday} and \code{7 == Saturday}.
\item \code{"iso"}: Encode weekdays assuming \code{1 == Monday} and \code{7 == Sunday}. This
is in line with the ISO standard.
}}
}
\value{
An integer vector of encodings.
}
\description{
\code{weekday_encoding()} extracts out an integer encoding for the weekday.
}
\examples{
# Here we supply a western encoding to `weekday()`
x <- weekday(1:7)
x

# We can extract out the codes using different encodings
weekday_encoding(x, encoding = "western")
weekday_encoding(x, encoding = "iso")
}
